- create dashboard module & page - setup dashboard to route - implement side menu with go router
