+++ lib/core/router/main_route.dart
@@ -1,14 +1,20 @@
-import '../../features/auth/presentations/pages/splash_page.dart';
+import '../../features/dashboard/presentation/page/dashboard_page.dart';
 import '../../features/main_page.dart';
 import '../core.dart';
 import 'auth_route.dart';
 import 'create_order_route.dart';
+import 'dashboard_route.dart';
 import 'process_order_route.dart';
 import 'profile_route.dart';
 
 export 'main_route.dart';
 
 final rootNavigatorKey = GlobalKey<NavigatorState>(debugLabel: "root");
+
+final dashboardRouteKey = GlobalKey<NavigatorState>(
+  debugLabel: "dashboard-route",
+);
+
 final createOrderRouteKey = GlobalKey<NavigatorState>(
   debugLabel: "create-order-route",
 );
@@ -19,7 +25,7 @@ final profilRouteKey = GlobalKey<NavigatorState>(debugLabel: "profile-route");
 
 final router = GoRouter(
   navigatorKey: rootNavigatorKey,
-  initialLocation: SplashPage.path,
+  initialLocation: DashboardPage.path,
   routes: [
     // auth route
     ...authRoute,
@@ -28,8 +34,11 @@ final router = GoRouter(
         return MainPage(navigationShell: navigationShell);
       },
       branches: [
-        // profile
-        StatefulShellBranch(navigatorKey: profilRouteKey, routes: profilRoute),
+        // dashboard
+        StatefulShellBranch(
+          navigatorKey: dashboardRouteKey,
+          routes: dashboardRoute,
+        ),
 
         // create order route
         StatefulShellBranch(
@@ -42,6 +51,15 @@ final router = GoRouter(
           navigatorKey: processOrderRoutKey,
           routes: processOrderRoute,
         ),
+
+        // inventaris
+        StatefulShellBranch(navigatorKey: profilRouteKey, routes: profilRoute),
+
+        // // notification
+        // StatefulShellBranch(navigatorKey: profilRouteKey, routes: profilRoute),
+
+        // // settings
+        // StatefulShellBranch(navigatorKey: profilRouteKey, routes: profilRoute),
       ],
     ),
   ],
